\name{ar1miss}
\alias{ar1miss}
\docType{data}
\title{AR with Missing Values
}
\description{Data used in Chapter 6
}
\format{
  The format is:  Time-Series [1:100]  with NA for missing values.
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}
