\name{arf}
\alias{arf}
\docType{data}
\title{Simulated ARFIMA
}
\description{1000 simulated observations from an ARFIMA(1, 1, 0) model with \eqn{\phi = .75} and \eqn{d = .4}.
}
\format{
  The format is:
 Time-Series [1:1000] from 1 to 1000: -0.0294 0.7487 -0.3386 -1.0332 -0.2627 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}
