\name{astsa-package}
\alias{astsa-package}
\alias{astsa}
\docType{package}
\title{Applied Statistical Time Series Analysis
}
\description{
Includes data and scripts to accompany \emph{Time Series Analysis and Its Applications: With R Examples} (4th ed) by R.H. Shumway and D.S. Stoffer. Springer Texts in Statistics, 2017. <DOI:10.1007/978-3-319-52452-8>.
}
\details{
\tabular{ll}{
Package: \tab astsa\cr
Type: \tab Package\cr
Version: \tab 1.8\cr
Date: \tab 2017-12-13\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
David Stoffer <stoffer@pitt.edu>
}
\references{See the webpage for the text: \url{http://www.stat.pitt.edu/stoffer/tsa4/} 
}
\note{This is most likely the last version of the package that I will submit to CRAN... I can NOT deal with these crazy people any more.
However, the most recent working version of the package will always be on GitHub.  See
\url{https://github.com/nickpoison/astsa} for details.
}
\keyword{ package }
\keyword{ ts }
