\name{beamd}
\alias{beamd}
\docType{data}
\title{Infrasonic Signal from a Nuclear Explosion 
}
\description{Infrasonic signal from a nuclear explosion.   
}
\usage{data(beamd)}
\format{
 A data frame with 2048 observations (rows) on 3 numeric variables (columns):  \code{sensor1}, \code{sensor2},  \code{sensor3}. 
}
\details{This is a data frame consisting of three columns (that are not time series objects). The data are an infrasonic signal from a nuclear explosion observed at sensors on a triangular array.
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} 
}
\keyword{datasets}
