\name{birth}
\alias{birth}
\docType{data}
\title{U.S. Monthly Live Births 
}
\description{Monthly live births (adjusted) in thousands for the United States, 1948-1979.
}
\format{
  The format is:
 Time-Series [1:373] from 1948 to 1979: 295 286 300 278 272 268 308 321 313 308 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}
