\name{bnrf1ebv}
\alias{bnrf1ebv}
\docType{data}
\title{Nucleotide sequence - BNRF1 Epstein-Barr 
}
\description{Nucleotide sequence of the BNRF1 gene of the Epstein-Barr virus
(EBV): 1=A, 2=C, 3=G, 4=T. The data are used in Chapter 7.
}
\format{
  The format is:
 Time-Series [1:3954] from 1 to 3954: 1 4 3 3 1 1 3 1 3 1 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
