\name{bnrf1hvs}
\alias{bnrf1hvs}
\docType{data}
\title{Nucleotide sequence - BNRF1 of Herpesvirus saimiri
}
\description{Nucleotide sequence of the BNRF1 gene of the herpesvirus saimiri (HVS): 1=A, 2=C, 3=G, 4=T. The data are used in Chapter 7.
}
\format{
  The format is:
 Time-Series [1:3741] from 1 to 3741: 1 4 3 2 4 4 3 4 4 4 ...
}
\keyword{datasets}
