\name{cardox}
\alias{cardox}
\docType{data}
\title{Monthly Carbon Dioxide Levels at Mauna Loa
}
\description{
Monthly mean carbon dioxide (in ppm) measured at Mauna Loa Observatory, Hawaii. 
This is an update to \code{co2} in the datasets package.
}
\format{
  The format is:
 Time-Series [1:729] from March, 1958 to November 2018: 315.71 317.45 317.50 317.10  ...
}
\details{
 The carbon dioxide data measured as the mole fraction in dry air, on Mauna Loa constitute the longest record of direct measurements of CO2 in the atmosphere. They were started by C. David Keeling of the Scripps Institution of Oceanography in March of 1958 at a facility of the National Oceanic and Atmospheric Administration. NOAA started its own CO2 measurements in May of 1974, and they have run in parallel with those made by Scripps since then. Data are reported as a dry mole fraction defined as the number of molecules of carbon dioxide divided by the number of molecules of dry air multiplied by one million (ppm). 
}
\source{\url{https://www.esrl.noaa.gov/gmd/ccgg/trends/}
}
\references{
\url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}
