\name{chicken}
\alias{chicken}
\docType{data}
\title{Monthly price of a pound of chicken
}
\description{
Poultry (chicken), Whole bird spot price, Georgia docks, US cents per pound
}
\usage{data("chicken")}
\format{
  The format is:
 Time-Series [1:180] from August 2001 to July 2016: 65.6 66.5 65.7 64.3 63.2 ...
}
\source{\url{https://www.indexmundi.com/commodities/}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
