\name{climhyd}
\alias{climhyd}
\docType{data}
\title{Lake Shasta inflow data
}
\description{Lake Shasta inflow data  This is a data frame.
}
\format{A data frame with 454 observations (rows) on the following 6 numeric variables (columns): \code{Temp}, \code{DewPt}, \code{CldCvr}, \code{WndSpd}, \code{Precip}, \code{Inflow}.
}
\details{The data are 454 months of measured values for the climatic variables: air temperature, dew point, cloud cover, wind speed, precipitation, and inflow, at Lake Shasta, California.  The man-made lake is famous for the placard stating, "We don't swim in your toilet, so don't pee in our lake."}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
