\name{cmort}
\alias{cmort}
\docType{data}
\title{Cardiovascular Mortality from the LA Pollution study
}
\description{Average weekly cardiovascular mortality in Los Angeles County; 508 six-day smoothed averages obtained by filtering daily values over the 10 year period 1970-1979.
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 97.8 104.6 94.4 98 95.8 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{lap}}}
\keyword{datasets}
