\name{cpg}
\alias{cpg}
\docType{data}
\title{Hard Drive Cost per GB 
}
\description{Median annual cost per gigabyte (GB) of storage.
}
\format{
  The format is:
 Time-Series [1:29] from 1980 to 2008:  213000.00 295000.00 260000.00 175000.00 160000.00   ...
}
\details{The median annual cost of hard drives used in computers. The data are retail prices per GB taken from a sample of manufacturers.   
}
\source{\url{http://ns1758.ca/winch/winchest.html}
}
\references{\url{http://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
