\name{econ5}
\alias{econ5}
\docType{data}
\title{Five Quarterly Economic Series
}
\description{Data frame containing quarterly U.S. unemployment, GNP, consumption, and government and private investment, from 1948-III to 1988-II.
}
\usage{data(econ5)}
\format{
  A data frame with 161 observations (rows) on the following 5 numeric variables (columns):  \code{unemp}, \code{gnp}, \code{consum}, \code{govinv},     \code{prinv}.
}
\source{Young, P.C. and Pedregal, D.J. (1999). Macro-economic relativity: government spending, private investment and unemployment in the USA 1948-1998. \emph{Structural Change and Economic Dynamics}, 10, 359-380.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.  
}
\keyword{datasets}