\name{EQ5}
\alias{EQ5}
\docType{data}
\title{Seismic Trace of Earthquake number 5 
}
\description{Seismic trace of an earthquake [two phases or arrivals along the surface, the primary wave (\eqn{t = 1,\dots,1024}) and the shear wave (\eqn{t = 1025,\dots,2048})] recorded at a seismic station.
}
\format{
  The format is:
 Time-Series [1:2048] from 1 to 2048: 0.01749 0.01139 0.01512 0.01477 0.00651 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{eqexp}}}
\keyword{datasets}
