\name{EQcount}
\alias{EQcount}
\docType{data}
\title{EQ Counts
}
\description{Series of annual counts of major earthquakes (magnitude 7 and above) in the
world between 1900 and 2006. }
\format{
  The format is:
 Time-Series [1:107] from 1900 to 2006: 13 14  8 10 16 26  ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.  
}
\source{Zucchini and MacDonald (2009). 
 Hidden Markov Models for Time Series: An Introduction using R.  CRC Press.
}
