\name{EXP6}
\alias{EXP6}
\docType{data}
\title{Seismic Trace of Explosion number 6
}
\description{Seismic trace of an explosion [two phases or arrivals along the surface, the primary wave (\eqn{t = 1,\dots,1024}) and the shear wave (\eqn{t = 1025,\dots,2048})] recorded at a seismic station.
}
\format{
  The format is:
 Time-Series [1:2048] from 1 to 2048: -0.001837 -0.000554 -0.002284 -0.000303 -0.000721 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{eqexp}}}
\keyword{datasets}
