\name{flu}
\alias{flu}
\docType{data}
\title{Monthly pneumonia and influenza deaths in the U.S., 1968 to 1978.
}
\description{Monthly pneumonia and influenza deaths per 10,000 people in the United States for 11 years, 1968 to 1978.
}
\usage{data(flu)}
\format{
  The format is:
 Time-Series [1:132] from 1968 to 1979: 0.811 0.446 0.342 0.277 0.248 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}