\name{gas}
\alias{gas}
\docType{data}
\title{Gas Prices
}
\description{New York Harbor conventional regular gasoline weekly spot price FOB (in cents per gallon) from 2000 to mid-2010.
}
\format{
  The format is:
 Time-Series [1:545] from 2000 to 2010: 70.6 71 68.5 65.1 67.9 ...
}
\details{Pairs with series \code{oil}
}
\source{Data were obtained from the URL: www.eia.doe.gov/dnav/pet/pet_pri_spt_s1_w.htm
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{oil}}}
\keyword{datasets}