\name{gdp}
\alias{gdp}
\docType{data}
\title{Quarterly U.S. GDP
}
\description{Seasonally adjusted quarterly U.S. GDP from 1947(1) to 2018(3).
}
\format{
  The format is:
 Time-Series [1:287] from 1947 to 2018: 2033 2028 2023 2055 2086 ...
}
\source{\url{https://tradingeconomics.com/united-states/gdp}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}