\name{gnp}
\alias{gnp}
\docType{data}
\title{Quarterly U.S. GNP
}
\description{Seasonally adjusted quarterly U.S. GNP from 1947(1) to 2002(3).
}
\format{
  The format is:
 Time-Series [1:223] from 1947 to 2002: 1489 1497 1500 1524 1547 ...
}
\source{\url{https://research.stlouisfed.org/}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\link{gdp}
}
\keyword{datasets}
