\name{gtemp_land}
\alias{gtemp_land}
\docType{data}
\title{Global mean land  temperature deviations - updated to 2017
}
\description{Annual temperature anomalies (in degress centigrade) averaged over the Earth's land area from 1880 to  2017. 
}
\format{
  The format is:
 Time-Series [1:138] from 1880 to 2017: -0.62 -0.45 -0.47 -0.62 -0.82 ...
}
\source{\url{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{http://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{gtemp_ocean}},   \code{\link{globtemp}}, \code{\link{globtempl}},  \code{\link{gtemp2}}}
\keyword{datasets}