\name{gtemp_ocean}
\alias{gtemp_ocean}
\docType{data}
\title{Global mean ocean temperature deviations - updated to 2017
}
\description{Annual sea surface temperature anomalies averaged over the part of the ocean that is free of ice at all times (open ocean) from 1880 to 2017.
}
\format{
  The format is:
 Time-Series [1:138] from 1880 to 2009:  -0.05  0.01  0.00 -0.06 -0.15 ...
}
\source{\url{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{gtemp_land}}, \code{\link{globtemp}}, \code{\link{globtempl}},  \code{\link{gtemp2}}}
\keyword{datasets}