\name{Hare}
\alias{Hare}
\docType{data}
\title{Snowshoe Hare
}
\description{This is one of the classic studies of predator-prey interactions, the 90-year data set is the number, in thousands, of snowshoe hare pelts  purchased by the Hudson's Bay Company of Canada. While this is an indirect measure of predation, the assumption is that there is a direct relationship between the number of pelts collected and the number of hare and lynx in the wild.
}
\usage{data("Hare")}
\format{
  The format is:
 Time-Series [1:91] from 1845 to 1935: 19.6 19.6 19.6 12 28 ...
}
\note{This data set pairs with \code{\link{Lynx}}.  
The data are in  units of one  thousand.
}
\source{From Odum's "Fundamentals of Ecology",  p. 191. 
Data listed at: people.whitman.edu/~hundledr/courses/M250F03/LynxHare.txt.
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{Lynx}}}
\keyword{datasets}
