\name{jj}
\alias{jj}
\docType{data}
\title{Johnson and Johnson Quarterly Earnings Per Share
}
\description{Johnson and Johnson quarterly earnings per share, 84 quarters (21 years) measured from the first quarter of 1960 to the last quarter of 1980.
}
\format{
  The format is:
 Time-Series [1:84] from 1960 to 1981: 0.71 0.63 0.85 0.44 0.61 0.69 0.92 0.55 0.72 0.77 ...
}
\details{This data set is also included with the R distribution as \code{JohnsonJohnson}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}