\name{lap}
\alias{lap}
\docType{data}
\title{LA Pollution-Mortality Study
}
\description{LA Pollution-Mortality Study (1970-1979, weekly data). 
}
\format{
  The format is:
  mts [1:508, 1:11] }
\details{
\tabular{ll}{
columns are time series       \tab   with names \cr  
 (1) Total Mortality          \tab   \code{tmort}\cr  
 (2) Respiratory Mortality    \tab   \code{rmort}\cr  
 (3) Cardiovascular Mortality \tab   \code{cmort}\cr  
 (4) Temperature              \tab   \code{tempr} \cr 
 (5) Relative Humidity        \tab   \code{rh}\cr   
 (6) Carbon Monoxide          \tab   \code{co}\cr  
 (7) Sulfur Dioxide           \tab   \code{so2}\cr  
 (8) Nitrogen Dioxide         \tab   \code{no2}\cr  
 (9) Hydrocarbons             \tab   \code{hycarb}\cr  
(10) Ozone                    \tab   \code{o3}\cr  
(11) Particulates             \tab   \code{part} \cr 
}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}