\name{lead}
\alias{lead}
\docType{data}
\title{Leading Indicator
}
\description{Leading indicator, 150 months; taken from Box and Jenkins (1970).  
}
\usage{data(lead)}
\format{
  The format is:
 Time-Series [1:150] from 1 to 150: 10.01 10.07 10.32 9.75 10.33 ...
}
\details{This is also the R time series \code{BJsales.lead}:  The sales time series \code{BJsales} and leading indicator \code{BJsales.lead} each contain 150 observations. The objects are of class "ts".
}
\seealso{\code{\link{sales}}}
\keyword{datasets}
