\name{Lynx}
\alias{Lynx}
\docType{data}
\title{Canadian Lynx
}
\description{This is one of the classic studies of predator-prey interactions, the 90-year data set is the number, in thousands, of  lynx  pelts  purchased by the Hudson's Bay Company of Canada. While this is an indirect measure of predation, the assumption is that there is a direct relationship between the number of pelts collected and the number of hare and lynx in the wild.
}
\usage{data("Lynx")}
\format{
  The format is:
 Time-Series [1:91] from 1845 to 1935: 30.1 45.1 49.1 39.5 21.2 ...
}
\note{The data are in  units of one  thousand.  This data set pairs with \code{\link{Hare}} and is NOT the same as  \code{\link{lynx}}.
}
\source{From Odum's "Fundamentals of Ecology",  p. 191.
http://people.whitman.edu/~hundledr/courses/M250F03/M250.html
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{Hare}}}
\keyword{datasets}
