\name{nyse}
\alias{nyse}
\docType{data}
\title{Returns of the New York Stock Exchange
}
\description{Returns of the New York Stock Exchange (NYSE) from February 2, 1984 to December 31, 1991.
}
\usage{data(nyse)}
\format{
  The format is:
 Time-Series [1:2000] from 1 to 2000: 0.00335 -0.01418 -0.01673 0.00229 -0.01692 ...
}
\source{S+GARCH module -  Version 1.1 Release 2: 1998
}
\keyword{datasets}
