\name{oil}
\alias{oil}
\docType{data}
\title{Crude oil, WTI spot price FOB
}
\description{Crude oil, WTI spot price FOB (in dollars per barrel), weekly data from 2000 to mid-2010.
}
\format{
  The format is:
 Time-Series [1:545] from 2000 to 2010: 26.2 26.1 26.3 24.9 26.3 ...
}
\details{pairs with the series \code{gas}
}
\source{Data were obtained from the URL: www.eia.doe.gov/dnav/pet/pet_pri_spt_s1_w.htm
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{gas}}}
\keyword{datasets}
