\name{part}
\alias{part}
\docType{data}
\title{Particulate levels from the LA pollution study
}
\description{Particulate series corresponding to \code{cmort} from the LA pollution study.
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 72.7 49.6 55.7 55.2 66 ...
}
\seealso{\code{\link{lap}}}
\keyword{datasets}
