\name{prodn}
\alias{prodn}
\docType{data}
\title{Monthly Federal Reserve Board Production Index
}
\description{Monthly Federal Reserve Board Production Index (1948-1978, n = 372 months).
}
\usage{data(prodn)}
\format{
  The format is:
 Time-Series [1:372] from 1948 to 1979: 40.6 41.1 40.5 40.1 40.4 41.2 39.3 41.6 42.3 43.2 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
