\name{qinfl}
\alias{qinfl}
\docType{data}
\title{Quarterly Inflation
}
\description{Quarterly inflation rate in the Consumer Price Index from 1953-Ito 1980-II, n = 110 observations.
}
\format{
  The format is:
 Time-Series [1:110] from 1953 to 1980: 1.673 3.173 0.492 -0.327 -0.333 ...
}
\details{pairs with qintr (interest rate)
}
\source{Newbold, P. and T. Bos (1985). \emph{Stochastic Parameter Regression Models}. Beverly Hills: Sage.
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{qintr}}}
\keyword{datasets}
