\name{qintr}
\alias{qintr}
\docType{data}
\title{Quarterly Interest Rate
}
\description{Quarterly interest rate recorded for Treasury bills from 1953-Ito 1980-II, n = 110 observations. 
}
\format{
  The format is:
 Time-Series [1:110] from 1953 to 1980: 1.98 2.15 1.96 1.47 1.06 ...
}
\details{pairs with qinfl (inflation)
}
\source{Newbold, P. and T. Bos (1985). \emph{Stochastic Parameter Regression Models}. Beverly Hills: Sage.
}
\references{See \url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{qinfl}}}
\keyword{datasets}
