\name{rec}
\alias{rec}
\docType{data}
\title{Recruitment (number of new fish index)
}
\description{Recruitment (index of the number of new fish) for a period of 453 months ranging over the years 1950-1987. Recruitment is loosely defined as an indicator of new members of a population to the first life stage at which natural mortality stabilizes near adult levels.
}
\usage{data(rec)}
\format{
  The format is:
 Time-Series [1:453] from 1950 to 1988: 68.6 68.6 68.6 68.6 68.6 ...
}
\details{can pair with \code{soi} (Southern Oscillation Index)
}
\source{Data furnished by Dr. Roy Mendelssohn of the Pacific Fisheries Environmental Laboratory, NOAA (personal communication). Further discussion of the concept of Recruitment may be found here: derekogle.com/fishR/examples/oldFishRVignettes/StockRecruit.pdf
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{soi}}}
\keyword{datasets}