\name{sales}
\alias{sales}
\docType{data}
\title{Sales
}
\description{Sales, 150 months; taken from Box and Jenkins (1970).
}
\format{
  The format is:
 Time-Series [1:150] from 1 to 150: 200 200 199 199 199 ...
}
\details{This is also the R data set \code{BJsales}: The sales time series \code{BJsales} and leading indicator \code{BJsales.lead} each contain 150 observations. The objects are of class "ts".
}
\seealso{\code{\link{lead}}}
\keyword{datasets}
