\name{salmon}
\alias{salmon}
\docType{data}
\title{
Monthly export price of salmon
}
\description{
Farm Bred Norwegian Salmon, export price, US Dollars per Kilogram
}
\format{
  The format is:
 Time-Series [1:166] from September 2003 to June 2017: 2.88 3.16 2.96 3.12 3.23 3.32 3.45 3.61 3.48 3.21 ...
}
\source{\url{https://www.indexmundi.com/commodities/}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsda/}
}
\keyword{datasets}
