\name{salt}
\alias{salt}
\docType{data}
\title{Salt Profiles
}
\description{Salt profiles taken over a spatial grid set out on an agricultural field, 64 rows at 17-ft spacing.
}
\usage{data(salt)}
\format{
  The format is:
 Time-Series [1:64] from 1 to 64: 6 6 6 3 3 3 4 4 4 1.5 ...
}
\details{pairs with \code{saltemp}, temperature profiles on the same grid
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{saltemp}}}
\keyword{datasets}
