\name{saltemp}
\alias{saltemp}
\docType{data}
\title{Temperature Profiles 
}
\description{Temperature profiles over a spatial grid set out on an agricultural field, 64 rows at 17-ft spacing.
}
\usage{data(saltemp)}
\format{
  The format is:
 Time-Series [1:64] from 1 to 64: 5.98 6.54 6.78 6.34 6.96 6.51 6.72 7.44 7.74 6.85 ...
}
\details{pairs with \code{salt}, salt profiles on the same grid
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{salt}}}
\keyword{datasets}
