\name{sarima.sim}
\alias{sarima.sim}
\title{ARIMA Simulation
}
\description{Simulate data from (seasonal) ARIMA models.
}
\usage{
sarima.sim(ar = NULL, d = 0, ma = NULL, sar = NULL, D = 0, sma = NULL, S = NULL, n = 500, rand.gen = rnorm, ...)
}
\arguments{
  \item{ar}{
coefficients of AR component (does not have to be specified if component not in the model)
}
  \item{d}{
order of regular difference (does not have to be specified if component not in the model)
}
  \item{ma}{
coefficients of MA component (does not have to be specified if component not in the model)
}
  \item{sar}{
coefficients of SAR component (does not have to be specified if component not in the model)
}
  \item{D}{
order of seasonal difference (does not have to be specified if component not in the model)
}
  \item{sma}{
coefficients of SMA component (does not have to be specified if component not in the model)
}
  \item{S}{
seasonal period (does not have to be specified if component not in the model)
}
  \item{n}{
desired sample size
}
  \item{rand.gen}{
optional; a function to generate the innovations (defaults to normal)
}
  \item{\dots}{
additional arguments for \code{rand.gen}. Most usefully, the standard deviation of the innovations generated by \code{rnorm} can be specified by \code{sd} or the mean by \code{mean} for adding drift in an integrated model (see details).
}
}
\details{
Will generate a time series of length n (defaults to 500) from the specified SARIMA model using simplified input. 
The use of the term \code{mean} refers to the generation of normal innovations.  For example,
\code{sarima.sim(ar=.9, mean=5)} will generate data using N(5,1) or 5+N(0,1) innovations, so that
the constant in the model is 5 and the mean of the AR model is 5/(1-.9) = 50. An simpler way to generate
a non-zero mean is to add it to the end; see the examples.
}
\value{
{a time series of length n from the specified SARIMA model with the specified frequency if the model is seasonal}
}
\references{\url{http://www.stat.pitt.edu/stoffer/tsa4/} and \url{http://www.stat.pitt.edu/stoffer/tsda/} 
}
\author{D.S. Stoffer
}
\note{The script uses \code{polyMul} to obtain the appropriate autoregressive and moving average polynomials from the specified model and feeds those into \code{.zarima_sim}.  The model autoregressive polynomial ('AR side' = AR x SAR) is checked for causality and the model moving average polynomial ('MA side' = MA x SMA) is checked invertibility.  The script stops and reports an error at the first violation of causality or invertibility; i.e., it will not report multiple errors.  
}
\examples{
## AR(1) with mean 50 [n = 500 is default]
y = sarima.sim(ar=.9) + 50
## ARIMA(0,1,1) with drift  
tsplot(sarima.sim(ma=-.8, d=1, mean=.1))
## SAR(1) example from text
Months = c("J","F","M","A","M","J","J","A","S","O","N","D")
sAR = sarima.sim(sar=.9, S=12, n=36)
tsplot(sAR, type='c')
points(sAR, pch=Months, cex=1.1, font=4, col=1:4)
## SARIMA(0,1,1)x(0,1,1)_12 - B&J's favorite
tsplot(sarima.sim(d=1, ma=-.4, D=1, sma=-.6, S=12, n=120))  
## infinite variance t-errors 
tsplot(sarima.sim(ar=.9, rand.gen=function(n, ...){rt(n, df=2)}))
}
\keyword{ ~ts}

