\name{so2}
\alias{so2}
\docType{data}
\title{SO2 levels from the LA pollution study
}
\description{Sulfur dioxide levels from the LA pollution study
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 3.37 2.59 3.29 3.04 3.39 2.57 2.35 3.38 1.5 2.56 ...
}
\seealso{\code{\link{lap}}}
\keyword{datasets}
