\name{sp500w}
\alias{sp500w}
\docType{data}
\title{Weekly Growth Rate of the Standard and Poor's 500
}
\description{Weekly closing returns of the SP 500 from 2003 to September,  2012.
}
\format{An 'xts' object on 2003-01-03 to 2012-09-28; 
  Indexed by objects of class: [Date] TZ: UTC
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
