\name{speech}
\alias{speech}
\docType{data}
\title{Speech Recording
}
\description{A small .1 second (1000 points) sample of recorded speech for the phrase "aaa...hhh".
}
\format{
  The format is:
 Time-Series [1:1020] from 1 to 1020: 1814 1556 1442 1416 1352 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
