\name{star}
\alias{star}
\docType{data}
\title{Variable Star 
}
\description{The  magnitude of a star taken at midnight for
600 consecutive days. The data are taken from the classic text,
The Calculus of Observations, a Treatise on Numerical Mathematics, by E.T. Whittaker and G. Robinson, 
(1923, Blackie and Son, Ltd.). 
}
\format{
  The format is:
 Time-Series [1:600] from 1 to 600: 25 28 31 32 33 33 32 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.  
}
\keyword{datasets}
