\name{sunspotz}
\alias{sunspotz}
\docType{data}
\title{Biannual Sunspot Numbers
}
\description{Biannual smoothed (12-month moving average) number of sunspots from June 1749 to December 1978; n = 459. The "z" on the end is to distinguish this series from the one included with R (called \code{sunspots}).
}
\format{
The format is: 
Time Series:
Start = c(1749, 1) 
End = c(1978, 1) 
Frequency = 2 
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}
}
\keyword{datasets}
