\name{tempr}
\alias{tempr}
\docType{data}
\title{Temperatures from the LA pollution study 
}
\description{Temperature series corresponding to \code{cmort} from the LA pollution study.
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 72.4 67.2 62.9 72.5 74.2 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.  
}
\seealso{\code{\link{lap}}}
\keyword{datasets}
