\name{unemp}
\alias{unemp}
\docType{data}
\title{U.S. Unemployment
}
\description{Monthly U.S. Unemployment series (1948-1978, n = 372)
}
\usage{data(unemp)}
\format{
  The format is:
 Time-Series [1:372] from 1948 to 1979: 235 281 265 241 201 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{UnempRate}}}
\keyword{datasets}
