\name{UnempRate}
\alias{UnempRate}
\docType{data}
\title{U.S. Unemployment Rate
}
\description{Monthly U.S. unemployment rate in percent unemployed (Jan, 1948 - Nov, 2016, n = 827)
}
\format{
  The format is:
 Time-Series [1:827] from c(1948,1) to c(2016,11): 4.0 4.7 4.5 4.0 3.4 3.9  ...
}
\source{\url{https://data.bls.gov/timeseries/LNU04000000}
}
\references{\url{http://www.stat.pitt.edu/stoffer/tsa4/}
}
%%\examples{
%%data(unemp)
%%plot(unemp)
%%## maybe str(unemp) ; plot(unemp) ...
%%}
\keyword{datasets}
%%