\name{varve}
\alias{varve}
\docType{data}
\title{Annual Varve Series
}
\description{Sedimentary deposits from one location in Massachusetts for 634 years, beginning nearly 12,000 years ago.
}
\format{
  The format is:
 Time-Series [1:634] from 1 to 634: 26.3 27.4 42.3 58.3 20.6 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\keyword{datasets}
