\name{WBC}
\alias{WBC}
\docType{data}
\title{White Blood Cell Levels
}
\description{WBC: Measurements made for 91 days on the three variables, log(white blood count) [WBC], log(platelet) [PLT] and hematocrit [HCT].   Missing data code is 0 (zero).
}
\format{
  The format is:
 Time-Series [1:91] from 1 to 91: 2.33 1.89 2.08 1.82 1.82 ...
}
\details{See Examples 6.1 amd 6.9 for more details.
}
\source{Jones, R.H. (1984). Fitting multivariate models to unequally spaced data. In \emph{Time Series Analysis of Irregularly Observed Data}, pp. 158-188. E. Parzen, ed. Lecture Notes in Statistics, 25, New York: Springer-Verlag.
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\seealso{\code{\link{blood}}, \code{\link{HCT}}, \code{\link{PLT}}}
\keyword{datasets}
