\name{zarima_sim}
\alias{zarima_sim}
\title{Support Script for ARIMA Simulation}
\description{
 This is \code{stats::arima.sim} with some improvements in error reporting.
 It is here only to be used by \code{sarima.sim}.
}
\usage{
arima_sim(model, n, rand.gen = rnorm, innov = rand.gen(n, \dots),
          n.start = NA, start.innov = rand.gen(n.start, \dots),
          \dots)
}
\arguments{
  \item{model}{A list with component \code{ar} and/or \code{ma} giving
    the AR and MA coefficients respectively.  Optionally a component
    \code{order} can be used.  An empty list gives an ARIMA(0, 0, 0)
    model, that is white noise.}
  \item{n}{length of output series, before un-differencing.  A strictly
    positive integer.}
  \item{rand.gen}{optional: a function to generate the innovations.}
  \item{innov}{an optional times series of innovations.  If not
    provided, \code{rand.gen} is used.}
  \item{n.start}{length of \sQuote{burn-in} period.  If \code{NA}, the
    default, a reasonable value is computed.}
  \item{start.innov}{an optional times series of innovations to be used
    for the burn-in period.  If supplied there must be at least
    \code{n.start} values (and \code{n.start} is by default computed
    inside the function).}
  \item{\dots}{additional arguments for \code{rand.gen}.  Most usefully,
    the standard deviation of the innovations generated by \code{rnorm}
    can be specified by \code{sd}.}
}
\details{
  \code{sarima.sim} uses this script to generate (seasonal) ARIMA models.  
  The model is checked for causality and invertibility.  This is an improved
  version of \code{stats::arima.sim}.
}
\value{
  A time-series object of class \code{"ts"}.
}
\notes{\code{stats::arima.sim} reported a stationarity error for stationary 
ARIMA models and did not check for invertibility. We could not perpetuate 
the misinformation. Also, \code{stats::arima.sim} only simulated from 
nonseasonal models.  Now you can use  \code{\link{sarima.sim}}, which calls
this script.
}
\seealso{
  \code{\link{sarima.sim}}
}
\keyword{ts}