\name{astsa-package}
\alias{astsa-package}
\alias{astsa}
\docType{package}
\title{Applied Statistical Time Series Analysis
}
\description{
Includes data and scripts to accompany \emph{Time Series Analysis and Its Applications: With R Examples} (3rd ed) by R.H. Shumway and D.S. Stoffer. Springer Texts in Statistics, 2011.
}
\details{
\tabular{ll}{
Package: \tab astsa\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-07-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
David Stoffer <stoffer@pitt.edu>
}
\references{See the webpage for the text: \url{http://www.stat.pitt.edu/stoffer/tsa3/} or its mirror \url{http://lib.stat.cmu.edu/general/stoffer/tsa3/}
}
\keyword{ package }
\keyword{ ts }
