\name{astsa-package}
\alias{astsa-package}
\alias{astsa}
\docType{package}
\title{Applied Statistical Time Series Analysis
}
\description{
Includes data and scripts to accompany \emph{Time Series Analysis and Its Applications: With R Examples} (4th ed) by R.H. Shumway and D.S. Stoffer. Springer Texts in Statistics, 2017. Softcover ISBN: 978-3-319-52451-1; eBook ISBN 978-3-319-52452-8; DOI: 10.1007/978-3-319-52452-8, and \emph{Time Series:  A Data Analysis Approach Using R},  Chapman-Hall, 2019. ISBN: 9780367221096.
}
\details{
\tabular{ll}{
Package: \tab astsa\cr
Type: \tab Package\cr
Version: \tab 1.8.6\cr
Date: \tab 2019-01-23\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
David Stoffer <stoffer@pitt.edu>
}
\references{See the webpage for the texts: \url{http://www.stat.pitt.edu/stoffer/tsa4/} or \url{http://www.stat.pitt.edu/stoffer/tsda/}
}
\keyword{ package }
\keyword{ ts }
