\name{astsa-package}
\alias{astsa-package}
\alias{astsa}
\docType{package}
\title{Applied Statistical Time Series Analysis
}
\description{
Includes data and scripts to accompany \emph{Time Series Analysis and Its Applications: With R Examples} (4th ed) by R.H. Shumway and D.S. Stoffer. Springer Texts in Statistics, 2017. Softcover ISBN: 978-3319524511; eBook ISBN 978-3319524528; DOI: 10.1007/978-3319524528, and \emph{Time Series:  A Data Analysis Approach Using R},  Chapman-Hall, 2019. ISBN: 978-0367221096.
}
\details{
\tabular{ll}{
Package: \tab astsa\cr
Type: \tab Package\cr
Version: \tab 1.12\cr
Date: \tab 2020-12-20\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
David Stoffer <stoffer@pitt.edu>
}
\references{The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.
The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.
}
\keyword{ package }
\keyword{ ts }
