\name{cpg}
\alias{cpg}
\docType{data}
\title{Cost per GB
}
\description{Median annual cost per gigabyte (GB) of storage.
}
\format{
  The format is:
 Time-Series [1:29] from 1980 to 2008: -0.2 -0.11 -0.1 -0.2 -0.28 -0.31 -0.3 -0.33 -0.2 -0.11 ...
}
\details{The cost of hard drives used in computers for storing data in large quantities decreased rapidly for many years.  The data are the median annual retail prices per GB of hard drives taken from a sample of manufacturers.   
}
\source{\url{http://ns1758.ca/winch/winchest.html}
}
\references{\url{http://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
