\name{econ5}
\alias{econ5}
\docType{data}
\title{Five Quarterly Economic Series
}
\description{Data frame containing quarterly U.S. unemployment, GNP, consumption, and government and private investment, from 1948-III to 1988-II.
}
\usage{data(econ5)}
\format{
  A data frame with 161 observations (rows) on the following 5 numeric variables (columns):  \code{unemp}, \code{gnp}, \code{consum}, \code{govinv},     \code{prinv}.
}
\source{Young, P.C. and Pedregal, D.J. (1999). Macro-economic relativity: government spending, private investment and unemployment in the USA 1948-1998. \emph{Structural Change and Economic Dynamics}, 10, 359-380.
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}