\name{EQcount}
\alias{EQcount}
\docType{data}
\title{EQ Counts
}
\description{Series of annual counts of major earthquakes (magnitude 7 and above) in the
world between 1900 and 2006. }
\format{
  The format is:
 Time-Series [1:107] from 1900 to 2006: 13 14  8 10 16 26  ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\source{Zucchini and MacDonald (2009). 
 Hidden Markov Models for Time Series: An Introduction using R.  CRC Press.
}
