\name{sarima.sim}
\alias{sarima.sim}
\title{ARIMA Simulation
}
\description{Simulate data from (seasonal) ARIMA models.
}
\usage{
sarima.sim(ar = NULL, d = 0, ma = NULL, sar = NULL, D = 0, sma = NULL, S = NULL, n = 500, rand.gen = rnorm, ...)
}
\arguments{
  \item{ar}{
coefficients of AR component (does not have to be specified if component not in the model)
}
  \item{d}{
order of regular difference (does not have to be specified if component not in the model)
}
  \item{ma}{
coefficients of MA component (does not have to be specified if component not in the model)
}
  \item{sar}{
coefficients of SAR component (does not have to be specified if component not in the model)
}
  \item{D}{
order of seasonal difference (does not have to be specified if component not in the model)
}
  \item{sma}{
coefficients of SMA component (does not have to be specified if component not in the model)
}
  \item{S}{
seasonal period (does not have to be specified if component not in the model)
}
  \item{n}{
desired sample size
}
  \item{rand.gen}{
optional; a function to generate the innovations (defaults to normal)
}
  \item{\dots}{
additional arguments for \code{rand.gen}. Most usefully, the standard deviation of the innovations generated by \code{rnorm} can be specified by \code{sd} or \code{mean}
}
}
\details{
Will generate a time series of length n from the specified SARIMA model using simplified input. 
}
\value{
{a time series of length n from the specified SARIMA model with the specified frequency the model is seasonal}
}
\references{\url{http://www.stat.pitt.edu/stoffer/tsa4/} and \url{http://www.stat.pitt.edu/stoffer/tsda/} 
}
\author{D.S. Stoffer
}
\note{
The script uses \code{polyMul} to obtain the appropriate autoregessive and moving average polynomials
from the specified model and feeds those into \code{stats::arima.sim}. 
}
\examples{
## AR(1)
y = sarima.sim(ar=.9, n=100)
tsplot(y)
acf2(y)
## ARIMA(0,1,1) with drift  
y = sarima.sim(ma=-.8, d=1, n=100, mean=.1)
## SAR(1) example from text
set.seed(666)
Months = c("J","F","M","A","M","J","J","A","S","O","N","D")
sAR = sarima.sim(sar=.9, S=12, n=36)
tsplot(sAR, type='c')
points(sAR, pch=Months, cex=1.1, font=4, col=1:4)
## SARIMA(0,1,1)x(0,1,1)_12 the old standby
y = sarima.sim(d=1, ma=-.8, D=1, sma=.8, S=12, n=120) 
acf2(diff(y)) 
## t-errors 
y = sarima.sim(n=60, ar=c(0.9, -0.5), ma = c(-0.2, 0.2), sma=.9, S=12, 
               rand.gen = function(n, ...) sqrt(0.2) * rt(n, df = 5))
}str
\keyword{ ~ts}

