\name{soi}
\alias{soi}
\docType{data}
\title{Southern Oscillation Index
}
\description{Southern Oscillation Index (SOI) for a period of 453 months ranging over the years 1950-1987.
}
\format{
  The format is:
 Time-Series [1:453] from 1950 to 1988: 0.377 0.246 0.311 0.104 -0.016 0.235 0.137 0.191 -0.016 0.29 ...
}
\details{pairs with \code{rec} (Recruitment)
}
\source{Data furnished by Dr. Roy Mendelssohn of the Pacific Fisheries Environmental Laboratory, NOAA (personal communication).
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{rec}}}
\keyword{datasets}