\name{soiltemp}
\alias{soiltemp}
\docType{data}
\title{Spatial Grid of Surface Soil Temperatures
}
\description{A 64 by 36 matrix of surface soil temperatures.
}
\format{
  The format is:
 num [1:64, 1:36] 6.7 8.9 5 6.6 6.1 7 6.5 8.2 6.7 6.6 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/}
}
\keyword{datasets}
