\name{star}
\alias{star}
\docType{data}
\title{Variable Star 
}
\description{The  magnitude of a star taken at midnight for
600 consecutive days. The data are taken from the classic text,
The Calculus of Observations, a Treatise on Numerical Mathematics, by E.T. Whittaker and G. Robinson, 
(1923, Blackie and Son, Ltd.). 
}
\format{
  The format is:
 Time-Series [1:600] from 1 to 600: 25 28 31 32 33 33 32 ...
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}
}
\keyword{datasets}
