\name{tempr}
\alias{tempr}
\docType{data}
\title{Temperatures from the LA pollution study 
}
\description{Temperature series corresponding to \code{cmort} from the LA pollution study.
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 72.4 67.2 62.9 72.5 74.2 ...
}
\seealso{\code{\link{lap}}}
\keyword{datasets}
