\name{UnempRate}
\alias{UnempRate}
\docType{data}
\title{U.S. Unemployment Rate
}
\description{Monthly U.S. unemployment rate in percent unemployed (Jan, 1948 - Nov, 2016, n = 827)
}
\format{
  The format is:
 Time-Series [1:827] from 1948 to 2017: 4 4.7 4.5 4 3.4 3.9 3.9 3.6 3.4 2.9 ...
}
\source{\url{https://data.bls.gov/timeseries/LNU04000000/}
}
\references{\url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/} 
}
\seealso{\code{\link{unemp}}}
\keyword{datasets}
